+++ src/features/create-post/ul/create-button.tsx
@@ -1,0 +1,8 @@
+import React from 'react'
+import Button from '@/shared/ui/Button/Button'
+
+const CreateButton = () => {
+  return <div></div>
+}
+
+export default CreateButton
